Add basic mimpmap downscaling support
This is small patch to fix quality issues when scaling images down.
It'll create a number of mipmap/half-scaled versions of a given source
as required and use the next largest version as the source for bilinear
or bicubic scaling.

This uses the same downsampling functions as Skia, but similarly doesn't
run SIMD versions of these. As they are integer-based, these should be
reasonably quick.